probando

+++ trunk/Src/matrix.cpp
@@ -13,7 +13,7 @@ Matrix::Matrix(int rows, int columns)
     }
 }
 
-int Matrix::at(int r, int c)
+float Matrix::at(int r, int c)
 {
    return this->values[r][c];
 }
